Adds an App Engine wih GCM module.
Mostly follows instructions on https://github.com/GoogleCloudPlatform/gradle-appengine-templates/tree/master/GcmEndpoints, and upcoming updates will also mostly follow that. This only adds the module. It doesn't include any of the changes in section 2 onward, wiring up the app module to the backend.

This update includes two changes not mentioned in the guide:
- Reduces the default gms core dependency added by that process to only depend on the gcm library.
- Updates appengine plugin to 1.9.30.