Add support for syncing with AWS DynamoDB via a SyncAdapter.
The service connects to AWS and downloads all Reporting Instructions records. It saves them locally on disk, and it only updates the local copy if it detects changes. Nothing is done with the synced instructions yet.

Some GCM code was removed in the process, since we won't be using any of that for now.