Some tests for CheckInAlarmReceiver.
First test showed that parsing error behavior was bad, throwing an exception and
crashing the component.

Now, notify the user about parse errors and reschedule (because it was too hard to add an
invalid bit to Instructions).

Also moves downloading into a Downloader class so that the test can mock out the
downloading.