Tweaks parsing to fix #24
The fix for #13 involved checking for "GROUP", but now they've used that in a comment
inside the GROUPS REPORTING section, not at the end of it. So, I've made the check more
detailed, looking for exactly the strings that I know could terminate the reporting section:
"GROUPS ON STANDBY", "GROUPS STANDBY", and "GROUPS EXCUSED".

In the process, found and fixed a catastrophic backtracking issue in one of the matchers. Not
sure if it was really an issue, but regex101 complained, and it was an easy fix.

Also, gained the ability to parse instructions from files as test resources. So I can add more
examples just by adding a file and adding a test with minimal boilerplate.